fix(tests): refactor udc tests
