+++ __tests__/defaultProvider.test.ts
@@ -1,6 +1,6 @@
 import { BlockNumber, GetBlockResponse, stark } from '../src';
 import { toBN } from '../src/utils/number';
-import { getERC20DeployPayload, getTestProvider } from './fixtures';
+import { erc20ClassHash, getERC20DeployPayload, getTestProvider } from './fixtures';
 
 const { compileCalldata } = stark;
 
@@ -8,7 +8,7 @@ const testProvider = getTestProvider();
 
 describe('defaultProvider', () => {
   let exampleTransactionHash: string;
-  let exampleContractAddress: string;
+  let erc20ContractAddress: string;
 
   let exampleBlock: GetBlockResponse;
   let exampleBlockNumber: BlockNumber;
@@ -23,7 +23,7 @@ describe('defaultProvider', () => {
     );
     await testProvider.waitForTransaction(transaction_hash);
     exampleTransactionHash = transaction_hash;
-    exampleContractAddress = contract_address;
+    erc20ContractAddress = contract_address;
 
     exampleBlock = await testProvider.getBlock('latest');
     exampleBlockHash = exampleBlock.block_hash;
@@ -32,7 +32,7 @@ describe('defaultProvider', () => {
 
   describe('endpoints', () => {
     test('deployContract()', () => {
-      expect(exampleContractAddress).toBeTruthy();
+      expect(erc20ContractAddress).toBeTruthy();
       expect(exampleTransactionHash).toBeTruthy();
     });
 
@@ -75,43 +75,41 @@ describe('defaultProvider', () => {
     });
 
     test('getNonce()', async () => {
-      const nonce = await testProvider.getNonce(exampleContractAddress);
+      const nonce = await testProvider.getNonce(erc20ContractAddress);
       return expect(nonce).toEqual('0x0');
     });
 
     test('getClassAt(contractAddress, blockNumber="latest")', async () => {
-      const classResponse = await testProvider.getClassAt(exampleContractAddress);
+      const classResponse = await testProvider.getClassAt(erc20ContractAddress);
 
       expect(classResponse).toHaveProperty('program');
       expect(classResponse).toHaveProperty('entry_points_by_type');
     });
 
     // TODO see if feasible to split
-    describe('getClassHashAt & GetClass', () => {
+    describe(' GetClass', () => {
       test('responses', async () => {
-        const classHash = await testProvider.getClassHashAt(exampleContractAddress);
-        expect(typeof classHash).toBe('string');
-
-        const classResponse = await testProvider.getClass(classHash);
+        const classResponse = await testProvider.getClassByHash(erc20ClassHash);
         expect(classResponse).toHaveProperty('program');
         expect(classResponse).toHaveProperty('entry_points_by_type');
+        expect(classResponse).toHaveProperty('abi');
       });
     });
 
     describe('getStorageAt', () => {
       test('with "key" type of number', () => {
-        return expect(testProvider.getStorageAt(exampleContractAddress, 0)).resolves.not.toThrow();
+        return expect(testProvider.getStorageAt(erc20ContractAddress, 0)).resolves.not.toThrow();
       });
 
       test('"key" type of string', () => {
         return expect(
-          testProvider.getStorageAt(exampleContractAddress, '0x0')
+          testProvider.getStorageAt(erc20ContractAddress, '0x0')
         ).resolves.not.toThrow();
       });
 
       test('with "key" type of BN', () => {
         return expect(
-          testProvider.getStorageAt(exampleContractAddress, toBN('0x0'))
+          testProvider.getStorageAt(erc20ContractAddress, toBN('0x0'))
         ).resolves.not.toThrow();
       });
     });
@@ -135,7 +133,7 @@ describe('defaultProvider', () => {
       test('callContract()', () => {
         return expect(
           testProvider.callContract({
-            contractAddress: exampleContractAddress,
+            contractAddress: erc20ContractAddress,
             entrypoint: 'balanceOf',
             calldata: compileCalldata({
               user: '0x9ff64f4ab0e1fe88df4465ade98d1ea99d5732761c39279b8e1374fa943e9b',
@@ -148,7 +146,7 @@ describe('defaultProvider', () => {
         return expect(
           testProvider
             .callContract({
-              contractAddress: exampleContractAddress,
+              contractAddress: erc20ContractAddress,
               entrypoint: 'balanceOf',
               calldata: compileCalldata({
                 user: wallet,
@@ -163,7 +161,7 @@ describe('defaultProvider', () => {
       test('callContract() - gateway error', async () => {
         return expect(
           testProvider.callContract({
-            contractAddress: exampleContractAddress,
+            contractAddress: erc20ContractAddress,
             entrypoint: 'non_existent_entrypoint',
             calldata: compileCalldata({
               user: '0xdeadbeef',
